sort out the test record batch generating codes.
