+++ test-utils/src/string_gen.rs
@@ -1,3 +1,4 @@
+use crate::array_gen::StringArrayGenerator;
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -14,27 +15,14 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-//
-// use arrow::array::{ArrayRef, GenericStringArray, OffsetSizeTrait, RecordBatch, UInt32Array};
+
 use crate::stagger_batch;
-use arrow::array::{ArrayRef, GenericStringArray, OffsetSizeTrait, UInt32Array};
 use arrow::record_batch::RecordBatch;
 use rand::rngs::StdRng;
 use rand::{thread_rng, Rng, SeedableRng};
 
 /// Randomly generate strings
-pub struct StringBatchGenerator {
-    //// The maximum length of the strings
-    pub max_len: usize,
-    /// the total number of strings in the output
-    pub num_strings: usize,
-    /// The number of distinct strings in the columns
-    pub num_distinct_strings: usize,
-    /// The percentage of nulls in the columns
-    pub null_pct: f64,
-    /// Random number generator
-    pub rng: StdRng,
-}
+pub struct StringBatchGenerator(StringArrayGenerator);
 
 impl StringBatchGenerator {
     /// Make batches of random strings with a random length columns "a" and "b".
@@ -44,8 +32,8 @@ impl StringBatchGenerator {
     pub fn make_input_batches(&mut self) -> Vec<RecordBatch> {
         // use a random number generator to pick a random sized output
         let batch = RecordBatch::try_from_iter(vec![
-            ("a", self.gen_data::<i32>()),
-            ("b", self.gen_data::<i64>()),
+            ("a", self.0.gen_data::<i32>()),
+            ("b", self.0.gen_data::<i64>()),
         ])
         .unwrap();
         stagger_batch(batch)
@@ -57,41 +45,15 @@ impl StringBatchGenerator {
     /// if large is true, the array is a LargeStringArray
     pub fn make_sorted_input_batches(&mut self, large: bool) -> Vec<RecordBatch> {
         let array = if large {
-            self.gen_data::<i32>()
+            self.0.gen_data::<i32>()
         } else {
-            self.gen_data::<i64>()
+            self.0.gen_data::<i64>()
         };
 
         let array = arrow::compute::sort(&array, None).unwrap();
 
         let batch = RecordBatch::try_from_iter(vec![("a", array)]).unwrap();
         stagger_batch(batch)
-    }
-
-    /// Creates a StringArray or LargeStringArray with random strings according
-    /// to the parameters of the BatchGenerator
-    fn gen_data<O: OffsetSizeTrait>(&mut self) -> ArrayRef {
-        // table of strings from which to draw
-        let distinct_strings: GenericStringArray<O> = (0..self.num_distinct_strings)
-            .map(|_| Some(random_string(&mut self.rng, self.max_len)))
-            .collect();
-
-        // pick num_strings randomly from the distinct string table
-        let indicies: UInt32Array = (0..self.num_strings)
-            .map(|_| {
-                if self.rng.gen::<f64>() < self.null_pct {
-                    None
-                } else if self.num_distinct_strings > 1 {
-                    let range = 1..(self.num_distinct_strings as u32);
-                    Some(self.rng.gen_range(range))
-                } else {
-                    Some(0)
-                }
-            })
-            .collect();
-
-        let options = None;
-        arrow::compute::take(&distinct_strings, &indicies, options).unwrap()
     }
 
     /// Return an set of `BatchGenerator`s that cover a range of interesting
@@ -109,31 +71,15 @@ impl StringBatchGenerator {
                 } else {
                     num_strings
                 };
-                cases.push(StringBatchGenerator {
+                cases.push(StringBatchGenerator(StringArrayGenerator {
                     max_len,
                     num_strings,
                     num_distinct_strings,
                     null_pct,
                     rng: StdRng::from_seed(rng.gen()),
-                })
+                }))
             }
         }
         cases
-    }
-}
-
-/// Return a string of random characters of length 1..=max_len
-fn random_string(rng: &mut StdRng, max_len: usize) -> String {
-    // pick characters at random (not just ascii)
-    match max_len {
-        0 => "".to_string(),
-        1 => String::from(rng.gen::<char>()),
-        _ => {
-            let len = rng.gen_range(1..=max_len);
-            rng.sample_iter::<char, _>(rand::distributions::Standard)
-                .take(len)
-                .map(char::from)
-                .collect::<String>()
-        }
     }
 }
